Made subdaily test data available as a fixture through the pmodel unit test suite
